\name{Annot-class}
\docType{class}
\alias{Annot-class}
\alias{Annot}

\title{Class "Annot"}
\description{S4 class for storing Annot analysis results}
\section{Slots}{
The following slots are defined for \link{Annot} objects:
\describe{
    \item{\code{results}:}{The accumulated annotation results}
    \item{\code{probe}:}{The probe id}
    \item{\code{symbol}:}{The official gene symbol}
    \item{\code{entrezId}:}{The Entrez gene Identifier}
    \item{\code{ensemblId}:}{The ensembl ID as indicated by ensembl}
    \item{\code{map}:}{The cytoband locations of the gene}
    }
}

\author{Argiris Sakellariou
\email{a.sakellariou@gonkhosp.gr}}
\keyword{classes}
