\name{DataLD-class}
\docType{class}
\alias{DataLD-class}
\alias{DataLD}

\title{Class "DataLD"}
\description{S4 class for storing DataLD analysis results}
\section{Slots}{
The following slots are defined for \link{DataLD} objects:
    \describe{
        \item{\code{trainObj}:}{An eSet class object of a training set}
        \item{\code{valObj}:}{An eSet class object of a validation set}
    }
}

\author{Argiris Sakellariou
    \email{a.sakellariou@gonkhosp.gr}
}
\keyword{classes}
