\name{mAPKL-package}
\alias{mAPKL-package}
\docType{package}
\title{
    A hybrid feature selection method for gene expression data
}
\description{
A hybrid FS method (mAP-KL), which combines multiple hypothesis testing and
affinity propagation (AP) clustering algorithm along with the Krzanowski &
Lai cluster quality index, to select an informative subset of genes.
}
\details{
    \tabular{ll}{
    Package: \tab mAPKL\cr
    Type: \tab Package\cr
    Version: \tab 0.99.01\cr
    Date: \tab 2014-05-07\cr
    License: \tab GPL (>= 2)\cr
    }
}
\author{
    Argiris Sakellariou
    Argiris Sakellariou <a.sakellariou@gonkhosp.gr>
}
\references{
A. Sakellariou, D. Sanoudou, and G. Spyrou, "Combining multiple hypothesis
testing and affinity propagation clustering leads to accurate, robust and
sample size independent classification on gene expression data,
" BMC Bioinformatics, vol. 13, p. 270, 2012.
}

\keyword{ htest }
